+++ plugin/Controllers/MainController.php
@@ -15,6 +15,7 @@ use GeminiLabs\SiteReviews\Commands\RegisterPointers;
 use GeminiLabs\SiteReviews\Commands\RegisterPostType;
 use GeminiLabs\SiteReviews\Commands\RegisterShortcodeButtons;
 use GeminiLabs\SiteReviews\Commands\RegisterShortcodes;
+use GeminiLabs\SiteReviews\Commands\RegisterTaxonomy;
 use GeminiLabs\SiteReviews\Commands\RegisterWidgets;
 use GeminiLabs\SiteReviews\Controllers\BaseController;
 use GeminiLabs\SiteReviews\Strings;
@@ -98,7 +99,7 @@ class MainController extends BaseController
 	 */
 	public function registerDashboardGlanceItems( array $items )
 	{
-		$post_type = 'site-review';
+		$post_type = $this->app->post_type;
 		$num_posts = wp_count_posts( $post_type );
 
 		if( !isset( $num_posts->publish ) || !$num_posts->publish ) {
@@ -126,7 +127,7 @@ class MainController extends BaseController
 	{
 		global $menu, $typenow;
 
-		$post_type = 'site-review';
+		$post_type = $this->app->post_type;
 
 		foreach( $menu as $key => $value ) {
 			if( !isset( $value[2] ) || $value[2] !== "edit.php?post_type={$post_type}" )continue;
@@ -147,12 +148,16 @@ class MainController extends BaseController
 	}
 
 	/**
+	 * @param string $post_type
+	 *
 	 * @return void
 	 *
 	 * @action add_meta_boxes_review
 	 */
-	public function registerMetaBox()
+	public function registerMetaBox( $post_type )
 	{
+		if( $post_type != $this->app->post_type )return;
+
 		add_meta_box( "{$this->app->id}_review", __( 'Details', 'site-reviews' ), [ $this, 'renderMetaBox'], null, 'side' );
 	}
 
@@ -165,7 +170,7 @@ class MainController extends BaseController
 	{
 		$command = new RegisterPointers([[
 			'id'       => 'glsr-pointer-pinned',
-			'screen'   => 'site-review',
+			'screen'   => $this->app->post_type,
 			'target'   => '#misc-pub-pinned',
 			'title'    => __( 'Pin Your Reviews', 'site-reviews' ),
 			'content'  => __( 'You can pin exceptional reviews so that they are always shown first in your widgets and shortcodes.', 'site-reviews' ),
@@ -188,7 +193,6 @@ class MainController extends BaseController
 		if( !$this->app->hasPermission() )return;
 
 		$command = new RegisterPostType([
-			'post_type'   => 'site-review',
 			'slug'        => 'reviews',
 			'single'      => __( 'Review', 'site-reviews' ),
 			'plural'      => __( 'Reviews', 'site-reviews' ),
@@ -197,14 +201,14 @@ class MainController extends BaseController
 			'public'      => false,
 			'has_archive' => false,
 			'show_ui'     => true,
-			'labels'      => (new Strings)->post_type_labels(),
+			'labels'      => glsr_resolve( 'Strings' )->post_type_labels(),
 			'columns'     => [
-				'title'  => '', // empty values use the default label
-				'author' => __( 'Author', 'site-reviews' ),
-				'site'   => __( 'Type', 'site-reviews' ),
-				'stars'  => __( 'Rating', 'site-reviews' ),
-				'sticky' => __( 'Pinned', 'site-reviews' ),
-				'date'   => '',
+				'title'    => '', // empty values use the default label
+				'category' => '',
+				'type'     => __( 'Type', 'site-reviews' ),
+				'stars'    => __( 'Rating', 'site-reviews' ),
+				'sticky'   => __( 'Pinned', 'site-reviews' ),
+				'date'     => '',
 			],
 		]);
 
@@ -220,7 +224,7 @@ class MainController extends BaseController
 	 */
 	public function registerRowActions( array $actions, WP_Post $post )
 	{
-		if( $post->post_type !== 'site-review' || $post->post_status === 'trash' ) {
+		if( $post->post_type !== $this->app->post_type || $post->post_status === 'trash' ) {
 			return $actions;
 		}
 
@@ -329,8 +333,28 @@ class MainController extends BaseController
 
 			if( !is_callable( $callback ) )continue;
 
-			add_submenu_page( 'edit.php?post_type=site-review', $title, $title, 'customize', $slug, $callback );
+			add_submenu_page( sprintf( 'edit.php?post_type=%s', $this->app->post_type ), $title, $title, 'customize', $slug, $callback );
 		}
+	}
+
+	/**
+	 * @return void
+	 *
+	 * @action init
+	 */
+	public function registerTaxonomy()
+	{
+		if( !$this->app->hasPermission() )return;
+
+		$command = new RegisterTaxonomy([
+			'hierarchical'      => true,
+			'meta_box_cb'       => [ $this, 'renderTaxonomyMetabox' ],
+			'public'            => false,
+			'show_admin_column' => true,
+			'show_ui'           => true,
+		]);
+
+		$this->execute( $command );
 	}
 
 	/**
@@ -431,9 +455,9 @@ class MainController extends BaseController
 	 */
 	public function renderMetaBox( WP_Post $post )
 	{
-		if( $post->post_type != 'site-review' )return;
+		if( $post->post_type != $this->app->post_type )return;
 
-		$this->render( 'edit/meta', ['post' => $post ] );
+		$this->render( 'edit/metabox-details', ['post' => $post ] );
 	}
 
 	/**
@@ -445,7 +469,7 @@ class MainController extends BaseController
 	{
 		global $post;
 
-		if( $post->post_type != 'site-review' )return;
+		if( $post->post_type != $this->app->post_type )return;
 
 		$pinned = get_post_meta( $post->ID, 'pinned', true );
 
@@ -459,8 +483,8 @@ class MainController extends BaseController
 	 */
 	public function renderReview( WP_Post $post )
 	{
-		if( $post->post_type != 'site-review' )return;
-		if( post_type_supports( 'site-review', 'title' ) )return;
+		if( $post->post_type != $this->app->post_type )return;
+		if( post_type_supports( $this->app->post_type, 'title' ) )return;
 		if( get_post_meta( $post->ID, 'site_name', true ) == 'local' )return;
 
 		$this->render( 'edit/review', ['post' => $post ] );
@@ -473,8 +497,8 @@ class MainController extends BaseController
 	 */
 	public function renderReviewNotice( WP_Post $post )
 	{
-		if( $post->post_type != 'site-review' )return;
-		if( post_type_supports( 'site-review', 'title' ) )return;
+		if( $post->post_type != $this->app->post_type )return;
+		if( post_type_supports( $this->app->post_type, 'title' ) )return;
 
 		$type = get_post_meta( $post->ID, 'site_name', true );
 
@@ -505,6 +529,26 @@ class MainController extends BaseController
 			'licenses' => __( 'Licenses', 'site-reviews' ),
 		],[
 			'settings' => $this->app->getDefaults(),
+		]);
+	}
+
+	/**
+	 * register_taxonomy() 'meta_box_cb' callback
+	 *
+	 * @return void
+	 */
+	public function renderTaxonomyMetabox( $post, $box )
+	{
+		if( $post->post_type != $this->app->post_type )return;
+
+		$taxonomy = isset( $box['args']['taxonomy'] )
+			? $box['args']['taxonomy']
+			: $this->app->taxonomy;
+
+		$this->render( 'edit/metabox-categories', [
+			'post'     => $post,
+			'tax_name' => esc_attr( $taxonomy ),
+			'taxonomy' => get_taxonomy( $taxonomy ),
 		]);
 	}