+++ views/menu/status.php
@@ -29,7 +29,7 @@
 					<a href="<?= admin_url( 'edit.php?post_type=site-review&page=' . glsr_app()->id . "&tab=settings&section={$key}" ); ?>"><?= $title; ?></a>
 				</td>
 				<td class="total-fetched column-primary">
-					<a href="<?= admin_url( "edit.php?post_type=site-review&post_status=all&site_name={$key}" ); ?>"><?= $db->getReviewCount( $key ); ?></a>
+					<a href="<?= admin_url( "edit.php?post_type=site-review&post_status=all&site_name={$key}" ); ?>"><?= $db->getReviewCount( 'site_name', $key ); ?></a>
 					<button type="button" class="toggle-row"><span class="screen-reader-text">Show more details</span></button>
 				</td>
 				<td class="last-fetch" data-colname="<?= __( 'Last fetch', 'site-reviews' ); ?>">